:recycle: Move mixin to utils
